@@ -1,14 +1,13 @@
 package smart_home.smarthome
 
-import android.app.Activity
-import android.app.AlertDialog
 import android.content.Context
 import android.os.Bundle
 import android.os.Handler
-import android.support.v4.app.Fragment
+import android.os.Looper
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
@@ -19,11 +18,11 @@ import java.util.Date
 import smart_home.smarthome.entities.DateDeserializer
 import smart_home.smarthome.entities.SensorData
 
-abstract class SensorsFragment(private val mId: Int, url: String) : Fragment(), ISmartHomeData {
+abstract class SensorsFragment(private val mId: Int, protected val params: IGraphParameters?, private val  dataType: String) : Fragment(), ISmartHomeData {
     protected val mGson: Gson
-    private val mService = buildService(url)
+    private val mService = buildService()
 
-    private val mHandler = Handler()
+    private val mHandler = Handler(Looper.getMainLooper())
 
     init {
         val gsonBuilder = GsonBuilder()
@@ -31,9 +30,22 @@ abstract class SensorsFragment(private val mId: Int, url: String) : Fragment(), 
         mGson = gsonBuilder.create()
     }
 
-    private fun buildService(url: String): SmartHomeService<List<SensorData>> {
+    private fun buildService(): SmartHomeService<List<SensorData>> {
         return SmartHomeService<List<SensorData>>()
-                .setRequest("GET /sensor_data?${url}")
+                .setRequest(buildRequest())
+    }
+
+    fun buildRequest(): String {
+        var requestString = "GET /sensor_data?data_type=$dataType"
+        if (params != null) {
+            if (params.getData().mUsePeriod) {
+                requestString += "&period=" + params.getData().mPeriod
+            } else {
+                requestString += "&start=" + params.getData().getFromDate() + "&end=" + params.getData().getToDate()
+            }
+            requestString += "&maxPoints=200"
+        }
+        return requestString
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
@@ -42,15 +54,16 @@ abstract class SensorsFragment(private val mId: Int, url: String) : Fragment(), 
         return inflater.inflate(mId, container, false)
     }
 
-    override fun onAttach(context: Context?) {
+    override fun onAttach(context: Context) {
         super.onAttach(context)
         refresh()
     }
 
     override fun refresh() {
-        mService.execute(object : SmartHomeService.Callback<List<SensorData>> {
+        mService.setRequest(buildRequest())
+        mService.doInBackground(object : SmartHomeService.Callback<List<SensorData>> {
             override fun deserialize(response: String): List<SensorData> {
-                return mGson.fromJson<List<SensorData>>(response, object : TypeToken<List<SensorData>>() {}.type)
+                return mGson.fromJson(response, object : TypeToken<List<SensorData>>() {}.type)
             }
 
             override fun isString(): Boolean {
