@@ -2,11 +2,11 @@ package smart_home.smarthome
 
 import android.content.res.Resources
 import android.graphics.Color
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
 
 import smart_home.smarthome.entities.SensorData
 
